// backend/src/proto/RecipePackage/Recipe.ts
// Original file: ../proto/recipe.proto


export interface Recipe {
  'name'?: (string);
  'description'?: (string);
  'ingredients'?: (string);
  'steps'?: (string);
}

export interface Recipe__Output {
  'name'?: (string);
  'description'?: (string);
  'ingredients'?: (string);
  'steps'?: (string);
}